@@ -1,5 +1,5 @@
 {
-  "name": "uoa-karate",
+  "name": "vue3",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -8,23 +8,20 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
+    "core-js": "^3.6.5",
     "swiper": "^6.4.15",
     "vue": "^3.0.0-beta.1",
-    "vue-meta": "^2.3.3",
-    "vue-router": "^3.1.6"
+    "vue-router": "^4.0.0-alpha.6"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^4.3.1",
-    "@vue/cli-plugin-eslint": "^4.3.1",
+    "@vue/cli-plugin-babel": "~4.5.0",
+    "@vue/cli-plugin-eslint": "~4.5.0",
     "@vue/cli-plugin-router": "~4.5.0",
-    "@vue/cli-service": "^4.3.1",
+    "@vue/cli-service": "~4.5.0",
     "@vue/compiler-sfc": "^3.0.0-beta.1",
-    "@vue/eslint-config-prettier": "5",
     "babel-eslint": "^10.1.0",
-    "eslint": "5",
-    "eslint-plugin-prettier": "^3.1.3",
+    "eslint": "^6.7.2",
     "eslint-plugin-vue": "^7.0.0-alpha.0",
-    "prettier": "^2.0.5",
     "vue-cli-plugin-vue-next": "~0.1.4"
   },
   "eslintConfig": {
@@ -36,19 +33,14 @@
       "plugin:vue/vue3-essential",
       "eslint:recommended"
     ],
-    "rules": {},
     "parserOptions": {
       "parser": "babel-eslint"
-    }
-  },
-  "postcss": {
-    "plugins": {
-      "autoprefixer": {}
-    }
+    },
+    "rules": {}
   },
   "browserslist": [
     "> 1%",
     "last 2 versions",
-    "not ie <= 8"
+    "not dead"
   ]
 }
