@@ -6,17 +6,12 @@
     <Footer>
       <QuoteDisplay class="quote-display" />
       <div class="github">
-
-        This site's sourcecode is hosted on <a
-          href="https://github.com/matt-auckland/uoa-karate"
-          target="_blank"
-        >Github</a>
+        This site's sourcecode is hosted on
+        <a href="https://github.com/matt-auckland/uoa-karate" target="_blank">Github</a>
       </div>
     </Footer>
   </div>
-
 </template>
-
 
 <script>
 import Header from "@/components/Header.vue";
@@ -28,18 +23,18 @@ export default {
   components: {
     Header,
     QuoteDisplay,
-    SitewideBanner
+    SitewideBanner,
   },
-  metaInfo: {
-    title: "UoA Karate Club",
-    meta: [
-      {
-        name: "description",
-        content:
-          "University of Auckland Karate Club. Find info on training times, locations and more."
-      }
-    ]
-  }
+  // metaInfo: {
+  //   title: "UoA Karate Club",
+  //   meta: [
+  //     {
+  //       name: "description",
+  //       content:
+  //         "University of Auckland Karate Club. Find info on training times, locations and more.",
+  //     },
+  //   ],
+  // },
 };
 </script>
 
