@@ -75,15 +75,15 @@ import Utils from "@/libs/utils.js";
 
 export default {
   name: "TestPage",
-  metaInfo: {
-    title: "Karate Knowledge Tester",
-    meta: [
-      {
-        name: "description",
-        content: "Test your Karate Knowledge."
-      }
-    ]
-  },
+  // metaInfo: {
+  //   title: "Karate Knowledge Tester",
+  //   meta: [
+  //     {
+  //       name: "description",
+  //       content: "Test your Karate Knowledge."
+  //     }
+  //   ]
+  // },
   components: {
     QuestionComponent
   },
